alanan/requirements.txt
Flask==3.1.0
tensorflow==2.19.0
numpy==2.1.3
opencv-python==4.11.0.86
google-generativeai==0.8.4
python-dotenv==1.1.0
gunicorn==23.0.0